resturcture test utils
