+++ ibft/storage/stores.go
@@ -1,10 +1,23 @@
 package storage
 
 import (
-	"github.com/bloxapp/ssv-spec/types"
+	spectypes "github.com/bloxapp/ssv-spec/types"
+	forksprotocol "github.com/bloxapp/ssv/protocol/forks"
 	"github.com/bloxapp/ssv/protocol/v2/qbft/storage"
+	"github.com/bloxapp/ssv/storage/basedb"
+	"go.uber.org/zap"
 	"sync"
 )
+
+func NewStoresFromRoles(db basedb.IDb, logger *zap.Logger, roles ...spectypes.BeaconRole) *QBFTStores {
+	stores := NewStores()
+
+	for _, role := range roles {
+		stores.Add(role, New(db, logger, role.String(), forksprotocol.GenesisForkVersion))
+	}
+
+	return stores
+}
 
 // QBFTStores wraps sync map with cast functions to qbft store
 type QBFTStores struct {
@@ -18,7 +31,7 @@ func NewStores() *QBFTStores {
 }
 
 // Get store from sync map by role type
-func (qs *QBFTStores) Get(role types.BeaconRole) qbftstorage.QBFTStore {
+func (qs *QBFTStores) Get(role spectypes.BeaconRole) qbftstorage.QBFTStore {
 	s, ok := qs.m.Load(role)
 	if ok {
 		qbftStorage := s.(qbftstorage.QBFTStore)
@@ -28,6 +41,6 @@ func (qs *QBFTStores) Get(role types.BeaconRole) qbftstorage.QBFTStore {
 }
 
 // Add store to sync map by role as a key
-func (qs *QBFTStores) Add(role types.BeaconRole, store qbftstorage.QBFTStore) {
+func (qs *QBFTStores) Add(role spectypes.BeaconRole, store qbftstorage.QBFTStore) {
 	qs.m.Store(role, store)
 }
