+++ operator/validator/controller.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"crypto/rsa"
 	"encoding/hex"
-	"github.com/bloxapp/ssv/protocol/v2/qbft"
+	commons2 "github.com/bloxapp/ssv/protocol/v2/qbft"
 	"sync"
 	"time"
 
@@ -566,8 +566,8 @@ func setupRunners(ctx context.Context, logger *zap.Logger, options validator.Opt
 	}
 
 	domainType := types.GetDefaultDomain()
-	generateConfig := func(role spectypes.BeaconRole) *qbft.Config {
-		return &qbft.Config{
+	generateConfig := func(role spectypes.BeaconRole) *commons2.Config {
+		return &commons2.Config{
 			Signer:      options.Signer,
 			SigningPK:   options.SSVShare.ValidatorPubKey, // TODO right val?
 			Domain:      domainType,
