+++ protocol/v2/message/consensus_test.go
@@ -4,7 +4,7 @@ import (
 	"sort"
 	"testing"
 
-	testing2 "github.com/bloxapp/ssv/protocol/v2/testing"
+	qbfttesting "github.com/bloxapp/ssv/protocol/v2/testing"
 
 	specqbft "github.com/bloxapp/ssv-spec/qbft"
 	spectypes "github.com/bloxapp/ssv-spec/types"
@@ -13,12 +13,12 @@ import (
 
 func TestAggregateSorting(t *testing.T) {
 	uids := []spectypes.OperatorID{spectypes.OperatorID(1), spectypes.OperatorID(2), spectypes.OperatorID(3), spectypes.OperatorID(4)}
-	secretKeys, _ := testing2.GenerateBLSKeys(uids...)
+	secretKeys, _ := qbfttesting.GenerateBLSKeys(uids...)
 
 	identifier := []byte("pk")
 
 	generateSignedMsg := func(operatorId spectypes.OperatorID) *specqbft.SignedMessage {
-		return testing2.SignMsg(t, secretKeys, []spectypes.OperatorID{operatorId}, &specqbft.Message{
+		return qbfttesting.SignMsg(t, secretKeys, []spectypes.OperatorID{operatorId}, &specqbft.Message{
 			MsgType:    specqbft.CommitMsgType,
 			Height:     0,
 			Round:      1,
